Adding config for beam search and other fixes
